+++ backend/canisters/group/impl/src/timer_job_types.rs
@@ -17,7 +17,7 @@ pub enum TimerJob {
     MakeTransfer(MakeTransferJob),
     RemoveExpiredEvents(RemoveExpiredEventsJob),
     NotifyEscrowCanisterOfDeposit(NotifyEscrowCanisterOfDepositJob),
-    // NotifyUserOfP2PTradeCompleted(NotifyUserOfP2PTradeCompletedJob),
+    // NotifyUserOfP2PSwapCompleted(NotifyUserOfP2PSwapCompletedJob),
 }
 
 #[derive(Serialize, Deserialize, Clone)]
@@ -237,7 +237,7 @@ impl Job for NotifyEscrowCanisterOfDepositJob {
             {
                 Ok(escrow_canister::notify_deposit::Response::Success(_)) => {
                     mutate_state(|state| {
-                        state.data.chat.events.accept_p2p_trade(
+                        state.data.chat.events.accept_p2p_swap(
                             self.user_id,
                             self.thread_root_message_index,
                             self.message_id,
@@ -247,7 +247,7 @@ impl Job for NotifyEscrowCanisterOfDepositJob {
                     });
                 }
                 Ok(escrow_canister::notify_deposit::Response::OfferExpired) => mutate_state(|state| {
-                    state.data.chat.events.unreserve_p2p_trade(
+                    state.data.chat.events.unreserve_p2p_swap(
                         self.user_id,
                         self.thread_root_message_index,
                         self.message_id,