+++ backend/canisters/group/impl/src/updates/accept_p2p_swap.rs
@@ -2,23 +2,23 @@ use crate::activity_notifications::handle_activity_notification;
 use crate::timer_job_types::NotifyEscrowCanisterOfDepositJob;
 use crate::{mutate_state, run_regular_jobs, RuntimeState};
 use canister_tracing_macros::trace;
-use group_canister::accept_p2p_trade_offer::{Response::*, *};
+use group_canister::accept_p2p_swap::{Response::*, *};
 use ic_cdk_macros::update;
 use icrc_ledger_types::icrc1::transfer::TransferError;
-use types::{AcceptSwapSuccess, Chat, MessageId, MessageIndex, ReserveP2PTradeResult, UserId};
+use types::{AcceptSwapSuccess, Chat, MessageId, MessageIndex, UserId};
 
 #[update]
 #[trace]
-async fn accept_p2p_trade_offer(args: Args) -> Response {
+async fn accept_p2p_swap(args: Args) -> Response {
     run_regular_jobs();
 
-    let ReserveP2PTradeOfferResult { user_id, c2c_args } = match mutate_state(|state| reserve_p2p_trade_offer(&args, state)) {
+    let ReserveP2PSwapResult { user_id, c2c_args } = match mutate_state(|state| reserve_p2p_swap(&args, state)) {
         Ok(result) => result,
         Err(response) => return *response,
     };
 
-    let result = match user_canister_c2c_client::c2c_accept_p2p_trade_offer(user_id.into(), &c2c_args).await {
-        Ok(user_canister::c2c_accept_p2p_trade_offer::Response::Success(transaction_id)) => {
+    let result = match user_canister_c2c_client::c2c_accept_p2p_swap(user_id.into(), &c2c_args).await {
+        Ok(user_canister::c2c_accept_p2p_swap::Response::Success(transaction_id)) => {
             NotifyEscrowCanisterOfDepositJob::run(
                 user_id,
                 c2c_args.offer_id,
@@ -30,7 +30,7 @@ async fn accept_p2p_trade_offer(args: Args) -> Response {
                 token1_txn_in: transaction_id,
             })
         }
-        Ok(user_canister::c2c_accept_p2p_trade_offer::Response::TransferError(TransferError::InsufficientFunds { .. })) => {
+        Ok(user_canister::c2c_accept_p2p_swap::Response::TransferError(TransferError::InsufficientFunds { .. })) => {
             InsufficientFunds
         }
         Ok(response) => InternalError(format!("{response:?}")),
@@ -44,12 +44,12 @@ async fn accept_p2p_trade_offer(args: Args) -> Response {
     result
 }
 
-struct ReserveP2PTradeOfferResult {
+struct ReserveP2PSwapResult {
     user_id: UserId,
-    c2c_args: user_canister::c2c_accept_p2p_trade_offer::Args,
+    c2c_args: user_canister::c2c_accept_p2p_swap::Args,
 }
 
-fn reserve_p2p_trade_offer(args: &Args, state: &mut RuntimeState) -> Result<ReserveP2PTradeOfferResult, Box<Response>> {
+fn reserve_p2p_swap(args: &Args, state: &mut RuntimeState) -> Result<ReserveP2PSwapResult, Box<Response>> {
     if state.data.is_frozen() {
         return Err(Box::new(ChatFrozen));
     }
@@ -64,19 +64,19 @@ fn reserve_p2p_trade_offer(args: &Args, state: &mut RuntimeState) -> Result<Rese
         let min_visible_event_index = member.min_visible_event_index();
         let now = state.env.now();
 
-        match state.data.chat.events.reserve_p2p_trade(
+        match state.data.chat.events.reserve_p2p_swap(
             user_id,
             args.thread_root_message_index,
             args.message_id,
             min_visible_event_index,
             now,
         ) {
-            ReserveP2PTradeResult::Success(result) => {
+            types::ReserveP2PSwapResult::Success(result) => {
                 handle_activity_notification(state);
 
-                Ok(ReserveP2PTradeOfferResult {
+                Ok(ReserveP2PSwapResult {
                     user_id,
-                    c2c_args: user_canister::c2c_accept_p2p_trade_offer::Args {
+                    c2c_args: user_canister::c2c_accept_p2p_swap::Args {
                         offer_id: result.content.offer_id,
                         chat: Chat::Group(state.env.canister_id().into()),
                         created: result.created,
@@ -90,8 +90,8 @@ fn reserve_p2p_trade_offer(args: &Args, state: &mut RuntimeState) -> Result<Rese
                     },
                 })
             }
-            ReserveP2PTradeResult::Failure(status) => Err(Box::new(StatusError(status.into()))),
-            ReserveP2PTradeResult::OfferNotFound => Err(Box::new(OfferNotFound)),
+            types::ReserveP2PSwapResult::Failure(status) => Err(Box::new(StatusError(status.into()))),
+            types::ReserveP2PSwapResult::OfferNotFound => Err(Box::new(OfferNotFound)),
         }
     } else {
         Err(Box::new(UserNotInGroup))
@@ -103,5 +103,5 @@ fn rollback(user_id: UserId, thread_root_message_index: Option<MessageIndex>, me
         .data
         .chat
         .events
-        .unreserve_p2p_trade(user_id, thread_root_message_index, message_id, state.env.now());
+        .unreserve_p2p_swap(user_id, thread_root_message_index, message_id, state.env.now());
 }