+++ backend/canisters/user/api/src/updates/send_message_with_transfer_to_channel.rs
@@ -34,7 +34,7 @@ pub enum Response {
     TransferFailed(String),
     TransferCannotBeZero,
     TransferCannotBeToSelf,
-    P2PTradeSetUpFailed(String),
+    P2PSwapSetUpFailed(String),
     UserSuspended,
     CommunityFrozen,
     RulesNotAccepted,