+++ backend/canisters/user/c2c_client/src/lib.rs
@@ -2,7 +2,7 @@ use canister_client::{generate_c2c_call, generate_candid_c2c_call};
 use user_canister::*;
 
 // Updates
-generate_c2c_call!(c2c_accept_p2p_trade_offer);
+generate_c2c_call!(c2c_accept_p2p_swap);
 generate_c2c_call!(c2c_charge_user_account);
 generate_c2c_call!(c2c_delete_messages);
 generate_c2c_call!(c2c_edit_message);