+++ backend/canisters/user/impl/src/updates/mod.rs
@@ -1,10 +1,10 @@
-pub mod accept_p2p_trade_offer;
+pub mod accept_p2p_swap;
 pub mod add_hot_group_exclusions;
 pub mod add_reaction;
 pub mod approve_transfer;
 pub mod archive_unarchive_chats;
 pub mod block_user;
-pub mod c2c_accept_p2p_trade_offer;
+pub mod c2c_accept_p2p_swap;
 pub mod c2c_charge_user_account;
 pub mod c2c_delete_messages;
 pub mod c2c_edit_message;