+++ backend/libraries/group_chat_core/src/lib.rs
@@ -1674,6 +1674,7 @@ impl GroupChatCore {
     }
 
     fn merge_message_permissions(new: OptionalMessagePermissions, old: MessagePermissions) -> MessagePermissions {
+        let p2p_swap = new.p2p_swap.apply_to(old.p2p_swap);
         MessagePermissions {
             default: new.default.unwrap_or(old.default),
             text: new.text.apply_to(old.text),
@@ -1685,7 +1686,8 @@ impl GroupChatCore {
             crypto: new.crypto.apply_to(old.crypto),
             giphy: new.giphy.apply_to(old.giphy),
             prize: new.prize.apply_to(old.prize),
-            p2p_trade: new.p2p_trade.apply_to(old.p2p_trade),
+            p2p_trade: p2p_swap,
+            p2p_swap,
             custom: GroupChatCore::merge_custom_permissions(new.custom_updated, new.custom_deleted, old.custom),
         }
     }